+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased] - 2025-10-26
+### Changed
+- Invalid structure error message:
+  - Now shows all errors instead of the first detected one.
+
+
 ## [2.2.4] – 2025-10-09
 ### Fixed
 - Added .env to .vscodeignore: